requirements.txt
pandas
numpy
scipy
sklearn.model_selection
sklearn.metrics
sklearn.neighbors